fix: ajustes no código
