+++ src/protected/application/lib/MapasCulturais/Repositories/Registration.php
@@ -167,19 +167,4 @@ class Registration extends \MapasCulturais\Repository {
             )
             $where";
     }
-
-    public static function getFinancialReportsAccountability($registration_id)
-    {
-        $app = App::i();
-
-        $query = "SELECT * FROM file WHERE object_id = :object_id AND grp = :grp";
-        $params = [
-            "object_id" => $registration_id,
-            "grp" => "financial-report-accountability",
-        ];
-        $conn = $app->em->getConnection();
-        $result = $conn->fetchAllAssociative($query, $params);
-
-        return $result;
-    }
 }
