+++ src/protected/application/lib/modules/Diligence/Repositories/Diligence.php
@@ -166,4 +166,18 @@ class Diligence{
         return $entity->getMetadata('situacion_diligence');        
     }
 
+    public static function getFinancialReportsAccountability($registration_id)
+    {
+        $app = App::i();
+
+        $query = "SELECT * FROM file WHERE object_id = :object_id AND grp = :grp";
+        $params = [
+            "object_id" => $registration_id,
+            "grp" => "financial-report-accountability",
+        ];
+        $conn = $app->em->getConnection();
+        $result = $conn->fetchAllAssociative($query, $params);
+
+        return $result;
+    }
 }
