+++ src/protected/application/lib/modules/Diligence/assets/css/diligence/multi.css
@@ -230,3 +230,19 @@ input[type="checkbox"] {
     border-radius: 3px; 
     margin-bottom: 10px
  }
+
+ .financial-report-wrapper {
+    background-color: #F5F5F5;
+    padding: 15px;
+    border-radius: 5px;
+    margin-bottom: 20px;
+    position: relative;
+ }
+
+ .delete-financial-report-btn {
+    position: absolute;
+    top: 4px;
+    right: 5px;
+    background-color: #E0E0E0;
+    border-radius: 3px;
+ }
