+++ src/protected/application/lib/modules/Diligence/assets/js/multi/multi.js
@@ -1,3 +1,40 @@
+const refo = {
+    activeEventDeleteFinancialReport() {
+        $('[delete-financial-report]').on('click', (event) => {
+            Swal.fire({
+                title: "Excluir Relatório Financeiro?",
+                text: "Essa ação não poderá ser desfeita.",
+                showCancelButton: true,
+                cancelButtonText: "Cancelar",
+                confirmButtonText: "Confirmar",
+                reverseButtons: true
+            }).then(res => {
+                if (res.isConfirmed) {
+                    $.ajax({
+                        type: "POST",
+                        url: MapasCulturais.createUrl('refo', 'deleteFinancialReport'),
+                        data: {
+                            fileId: event.currentTarget.dataset.fileId
+                        },
+                        dataType: "json",
+                        success() {
+                            $(event.currentTarget).parents('#financial-report-wrapper').remove()
+                            MapasCulturais.Messages.success('O relatório financeiro foi removido da prestação de contas')
+                        },
+                        error(err) {
+                            if (err.status === 400) {
+                                MapasCulturais.Messages.alert('O arquivo não pode ser removido, pois o TADO já foi gerado')
+                                return
+                            }
+                            MapasCulturais.Messages.error('Ocorreu algum erro. Verifique e tente novamente.')
+                        }
+                    })
+                }
+            })
+        })
+    }
+}
+
 $(document).ready(function () {
     $("#p-btn-tado").hide();
     $(".multi-itens-select").hide();
@@ -10,6 +47,14 @@ $(document).ready(function () {
         $("#p-btn-tado").show();
         
     });
+
+    $('#import-financial-report .mc-submit').on('click', () => {
+        setTimeout(() => {
+            refo.activeEventDeleteFinancialReport()
+        }, 1500)
+    })
+
+    refo.activeEventDeleteFinancialReport()
 });
 
 /**
