+++ src/protected/application/lib/modules/Diligence/layouts/parts/diligence/body-diligence-multi.php
@@ -6,18 +6,17 @@ use Diligence\Entities\AnswerDiligence;
 use Carbon\Carbon;
 use Diligence\Entities\Tado;
 use Diligence\Repositories\Diligence as DiligenceRepo;
-use MapasCulturais\Repositories\Registration as RegistrationRepo;
 
 if ($diligenceAndAnswers) :
     if ($diligenceAndAnswers[0]->status == EntityDiligence::STATUS_SEND) : ?>
         <div>
             <div class="import-financial-report">
                 <?php
-                $financialReportsAccountability = RegistrationRepo::getFinancialReportsAccountability($entity->id);
+                $financialReportsAccountability = DiligenceRepo::getFinancialReportsAccountability($entity->id);
                 $generatedTado = DiligenceRepo::getTado($entity);
 
                 $delBtn = '
-                    <div style="position: absolute; top: 4px; right: 5px; background-color: #E0E0E0; border-radius: 3px;">
+                    <div class="delete-financial-report-btn">
                         <a delete-financial-report data-file-id="%s" class="icon icon-close hltip" title="Excluir arquivo">
                         </a>
                     </div>
@@ -28,7 +27,7 @@ if ($diligenceAndAnswers) :
                     foreach ($financialReportsAccountability as $financialReportAccountability) {
                         $file_id = $financialReportAccountability["id"];
                         echo '
-                            <div style="background-color: #F5F5F5; padding: 15px; border-radius: 5px; margin-bottom: 20px; position: relative;" id="financial-report-wrapper">
+                            <div class="financial-report-wrapper" id="financial-report-wrapper">
                                 <i class="fas fa-download" style="margin-right: 10px;"></i>
                                 <a href="/arquivos/privateFile/' . $file_id . '" target="_blank" rel="noopener noreferrer">
                                     relatorio_financeiro.pdf
@@ -52,7 +51,7 @@ if ($diligenceAndAnswers) :
                     </label>
                     <label for="">
                         <strong>Assunto(s): </strong>
-                        <?php  $diligenceAndAnswers > 0 ? $diligenceAndAnswers[0]->getSubject() : ""; ?>
+                        <?php $diligenceAndAnswers > 0 ? $diligenceAndAnswers[0]->getSubject() : ""; ?>
                     </label>
                     <p style="margin: 10px 0px;">
                         <?php echo $diligenceAndAnswers[0]->description; ?>
@@ -115,41 +114,13 @@ if ($diligenceAndAnswers) :
                 $dt = null;
                 $dtSend = "";
 
-            if ($resultsDiligence !== null) {
-                $dt             = Carbon::parse($resultsDiligence->sendDiligence);
-                $dtSend         = $dt->isoFormat('LLL');
-                
-            }
-            
-            if($key > 1) {
-                if ($resultsDiligence instanceof EntityDiligence && !is_null($resultsDiligence) && $resultsDiligence->status == EntityDiligence::STATUS_SEND) { ?>
-                    <div style="display: flex; justify-content: space-between;" class="div-accordion-diligence">
-                        <label style="font-size: 14px">
-                            <b>Diligência:</b>
-                        </label>
-                        <label style="color: #085E55; font-size: 14px" class="title-hide-show-accordion">Visualizar <i class="fas fa-angle-down arrow"></i></label>
-                    </div>
-                    <div class="content">
-                        <p>
-                            <label for="">
-                                <strong>Assunto(s): </strong>
-                                <?php echo $resultsDiligence->getSubject(); ?>
-                            </label>
-                        </p>
-                        <p>
-                            <?php
-                            echo $resultsDiligence->description;
-                            ?>
-                        </p>
-                        <p class="paragraph-createTimestamp paragraph_createTimestamp_answer">
-                            <?php echo $dtSend; ?>
-                        </p>
-                    </div>
-                <?php
+                if ($resultsDiligence !== null) {
+                    $dt             = Carbon::parse($resultsDiligence->sendDiligence);
+                    $dtSend         = $dt->isoFormat('LLL');
                 }
 
                 if ($key > 1) :
-                    if ($resultsDiligence instanceof EntityDiligence && !is_null($resultsDiligence) && $resultsDiligence->status == EntityDiligence::STATUS_SEND) : ?>
+                    if ($resultsDiligence instanceof EntityDiligence && !is_null($resultsDiligence) && $resultsDiligence->status == EntityDiligence::STATUS_SEND) { ?>
                         <div style="display: flex; justify-content: space-between;" class="div-accordion-diligence">
                             <label style="font-size: 14px">
                                 <b>Diligência:</b>
@@ -158,48 +129,76 @@ if ($diligenceAndAnswers) :
                         </div>
                         <div class="content">
                             <p>
-                                <?php echo $resultsDiligence->description; ?>
+                                <label for="">
+                                    <strong>Assunto(s): </strong>
+                                    <?php echo $resultsDiligence->getSubject(); ?>
+                                </label>
+                            </p>
+                            <p>
+                                <?php
+                                echo $resultsDiligence->description;
+                                ?>
                             </p>
                             <p class="paragraph-createTimestamp paragraph_createTimestamp_answer">
                                 <?php echo $dtSend; ?>
                             </p>
                         </div>
-                    <?php
-                    endif;
+                        <?php
+                    }
 
-                    if ($resultsDiligence instanceof AnswerDiligence && !is_null($resultsDiligence) && $resultsDiligence->status == AnswerDiligence::STATUS_SEND) :
-                        $dtAnswer = Carbon::parse($resultsDiligence->createTimestamp);
-                        $dtSendAnswer = $dtAnswer->isoFormat('LLL');
-                    ?>
-                        <div style="display: flex; justify-content: space-between;" class="div-accordion-diligence">
-                            <label style="font-size: 14px">
-                                <b>Resposta recebida:</b>
-                            </label>
-                            <label style="color: #085E55; font-size: 14px" class="title-hide-show-accordion">
-                                Visualizar <i class="fas fa-angle-down arrow"></i>
-                            </label>
-                        </div>
-                        <div class="content" style="font-size: 14px; background-color: #F5F5F5; padding: 10px;">
-                            <p style="margin: 10px 0px;">
-                                <?php echo $resultsDiligence->answer; ?>
-                            </p>
-                            <?php
-                            $files = DiligenceRepo::getFilesDiligence($resultsDiligence->diligence->id);
+                    if ($key > 1) :
+                        if ($resultsDiligence instanceof EntityDiligence && !is_null($resultsDiligence) && $resultsDiligence->status == EntityDiligence::STATUS_SEND) : ?>
+                            <div style="display: flex; justify-content: space-between;" class="div-accordion-diligence">
+                                <label style="font-size: 14px">
+                                    <b>Diligência:</b>
+                                </label>
+                                <label style="color: #085E55; font-size: 14px" class="title-hide-show-accordion">Visualizar <i class="fas fa-angle-down arrow"></i></label>
+                            </div>
+                            <div class="content">
+                                <p>
+                                    <?php echo $resultsDiligence->description; ?>
+                                </p>
+                                <p class="paragraph-createTimestamp paragraph_createTimestamp_answer">
+                                    <?php echo $dtSend; ?>
+                                </p>
+                            </div>
+                        <?php
+                        endif;
 
-                            foreach ($files as $file) {
-                                echo '
+                        if ($resultsDiligence instanceof AnswerDiligence && !is_null($resultsDiligence) && $resultsDiligence->status == AnswerDiligence::STATUS_SEND) :
+                            $dtAnswer = Carbon::parse($resultsDiligence->createTimestamp);
+                            $dtSendAnswer = $dtAnswer->isoFormat('LLL');
+                        ?>
+                            <div style="display: flex; justify-content: space-between;" class="div-accordion-diligence">
+                                <label style="font-size: 14px">
+                                    <b>Resposta recebida:</b>
+                                </label>
+                                <label style="color: #085E55; font-size: 14px" class="title-hide-show-accordion">
+                                    Visualizar <i class="fas fa-angle-down arrow"></i>
+                                </label>
+                            </div>
+                            <div class="content" style="font-size: 14px; background-color: #F5F5F5; padding: 10px;">
+                                <p style="margin: 10px 0px;">
+                                    <?php echo $resultsDiligence->answer; ?>
+                                </p>
+                                <?php
+                                $files = DiligenceRepo::getFilesDiligence($resultsDiligence->diligence->id);
+
+                                foreach ($files as $file) {
+                                    echo '
                                     <p style="margin-bottom: 10px;">
                                         <a href="/arquivos/privateFile/' . $file["id"] . '" target="_blank" rel="noopener noreferrer">
                                             ' . $file["name"] . '
                                         </a>
                                     </p>
                                 ';
-                            }
-                            ?>
-                            <span style="font-size: 12px; font-weight: 700; color: #404040;">
-                                <?php echo $dtSendAnswer; ?>
-                            </span>
-                        </div>
+                                }
+                                ?>
+                                <span style="font-size: 12px; font-weight: 700; color: #404040;">
+                                    <?php echo $dtSendAnswer; ?>
+                                </span>
+                            </div>
+                        <?php endif; ?>
                     <?php endif; ?>
                 <?php endif; ?>
             <?php endforeach; ?>
@@ -207,27 +206,30 @@ if ($diligenceAndAnswers) :
     <?php endif; ?>
 <?php endif; ?>
 
-    <div class="div-diligence" id="div-diligence">
-        <p id="paragraph_info_status_diligence"></p>
-        <?php
-        //Array para marcar como confirmado a opção
-        isset($diligenceAndAnswers[0]) ?  $subjectReplace = $diligenceAndAnswers[0]->subjectToArray() : $subjectReplace = [];
-        isset($diligenceAndAnswers[0]) ? $checked = $diligenceAndAnswers[0]->getCheckSubject($subjectReplace) : $checked = ['checkPhysical' => "checked", 'checkFinance' => ""];
-        //Se não tiver resposta não mostra o assunto
-       if(!is_null($diligenceAndAnswers[1]) || isset($diligenceAndAnswers[1]->status))
-       {
-            $this->part('diligence/body-diligence-subject',
+<div class="div-diligence" id="div-diligence">
+    <p id="paragraph_info_status_diligence"></p>
+    <?php
+    //Array para marcar como confirmado a opção
+    isset($diligenceAndAnswers[0]) ?  $subjectReplace = $diligenceAndAnswers[0]->subjectToArray() : $subjectReplace = [];
+    isset($diligenceAndAnswers[0]) ? $checked = $diligenceAndAnswers[0]->getCheckSubject($subjectReplace) : $checked = ['checkPhysical' => "checked", 'checkFinance' => ""];
+    //Se não tiver resposta não mostra o assunto
+    if (!is_null($diligenceAndAnswers[1]) || isset($diligenceAndAnswers[1]->status)) {
+        $this->part(
+            'diligence/body-diligence-subject',
             [
-                'checkPhysical' => $checked['checkPhysical'] ,
-                'checkFinance' => $checked['checkFinance'] ,
-            ]);
-       }else {
-           //Mostrará assunto quando não tiver diligencia ou quando editar rascunho
-           $this->part('diligence/body-diligence-subject',
-               [
-                   'checkPhysical' => $checked['checkPhysical'],
-                   'checkFinance' => $checked['checkFinance'],
-               ]);
-       }
-       ?>
-    </div>
+                'checkPhysical' => $checked['checkPhysical'],
+                'checkFinance' => $checked['checkFinance'],
+            ]
+        );
+    } else {
+        //Mostrará assunto quando não tiver diligencia ou quando editar rascunho
+        $this->part(
+            'diligence/body-diligence-subject',
+            [
+                'checkPhysical' => $checked['checkPhysical'],
+                'checkFinance' => $checked['checkFinance'],
+            ]
+        );
+    }
+    ?>
+</div>
