+++ src/protected/application/lib/modules/Diligence/layouts/parts/diligence/body-proponent-multi.php
@@ -5,7 +5,6 @@ use Diligence\Entities\Diligence as EntityDiligence;
 use Diligence\Entities\AnswerDiligence;
 use Carbon\Carbon;
 use Diligence\Repositories\Diligence as DiligenceRepo;
-use MapasCulturais\Repositories\Registration as RegistrationRepo;
 
 if ($diligenceAndAnswers) {
     $diligencesSent = array_filter($diligenceAndAnswers, function ($value, $key) use ($diligenceAndAnswers) {
@@ -27,13 +26,13 @@ if ($diligenceAndAnswers) {
 <?php if (isset($diligencesSentReindexed)) : ?>
     <div class="import-financial-report">
         <?php
-        $financialReportsAccountability = RegistrationRepo::getFinancialReportsAccountability($entity->id);
+        $financialReportsAccountability = DiligenceRepo::getFinancialReportsAccountability($entity->id);
 
         if ($financialReportsAccountability) {
             foreach ($financialReportsAccountability as $financialReportAccountability) {
                 $file_id = $financialReportAccountability["id"];
                 echo '
-                    <div style="background-color: #F5F5F5; padding: 15px; border-radius: 5px; margin-bottom: 20px; position: relative;" id="financial-report-wrapper">
+                    <div class="financial-report-wrapper" id="financial-report-wrapper">
                         <i class="fas fa-download" style="margin-right: 10px;"></i>
                         <a href="/arquivos/privateFile/' . $file_id . '" target="_blank" rel="noopener noreferrer">
                             relatorio_financeiro.pdf
